[Util] Fix error in using fp16.h functions
This PR fixes multiple definition error when using FP32/16 conversion functions in fp16.h

**Changes proposed in this PR:**
- Function definition is moved to fp16.cpp.

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped
